swap chain migrated and unmigrated fixtures
