+++ tests/pytest_plugin/test_unmigrated_chain_fixture.py
@@ -1,0 +1,8 @@
+def test_chain_fixture(request, project_dir):
+    chain = request.getfuncargvalue('unmigrated_chain')
+
+    web3 = chain.web3
+
+    client_name, _, _, _ = web3.version.node.split('/')
+
+    assert client_name == "TestRPC"
